perf: lazy-load heavy deps to speed up CLI startup
Importing mlipaudit.benchmarks (which the CLI does to build the argument
parser) eagerly pulled in mlip.models, JAX/JAX-MD and scikit-learn, making
even `mlipaudit -h` take ~6s. These libraries are only needed when a
benchmark is actually run.

Defer those imports to call time:
- benchmark.py / utils.{inference,simulation,stability,unallowed_elements}:
  move mlip.models / mlip.simulation.ase (JAX-MD) / jax / sklearn imports
  into the functions that use them; type-only refs go under TYPE_CHECKING
  with `from __future__ import annotations`.
- Move ASESimulationEngineWithCalculator into utils/_ase_engine.py so the
  heavy ASE-engine base class is only imported when a simulation runs; keep
  the historical import paths working via module-level __getattr__.
- nudged_elastic_band / conformer_selection / dihedral_scan /
  water_radial_distribution: defer their mlip / sklearn imports.
- main.py: import launch_app / run_benchmarks inside their command branches.

Tests patched the now-lazy symbols at the use-site module; update them to
patch at the definition site (e.g. mlip.simulation.jax_md.JaxMDSimulationEngine),
which the lazy `from ... import` picks up at call time.

`mlipaudit -h` drops from ~5.6s to ~0.9s; importing mlipaudit.benchmarks
drops from ~5.0s to ~1.1s. All 128 tests pass.
