+++ tests/conformer_selection/test_conformer_selection.py
@@ -75,7 +75,7 @@ def test_full_run_with_mocked_inference(
     benchmark = conformer_selection_benchmark
 
     _mocked_batched_inference = mocker.patch(
-        "mlipaudit.utils.inference.run_batched_inference",
+        "mlip.inference.run_batched_inference",
         side_effect=mocked_batched_inference,
     )
