+++ tests/dihedral_scan/test_dihederal_scan.py
@@ -59,7 +59,7 @@ def test_full_run_with_mocked_inference(
     benchmark = dihedral_scan_benchmark
 
     _mocked_batched_inference = mocker.patch(
-        "mlipaudit.utils.inference.run_batched_inference",
+        "mlip.inference.run_batched_inference",
         side_effect=mocked_batched_inference,
     )
