+++ tests/ring_planarity/test_ring_planarity.py
@@ -107,7 +107,7 @@ def test_full_run_with_mocked_engine(
     benchmark = ring_planarity_benchmark
     mock_engine = mock_jaxmd_simulation_engine()
     with patch(
-        "mlipaudit.utils.simulation.JaxMDSimulationEngine",
+        "mlip.simulation.jax_md.JaxMDSimulationEngine",
         return_value=mock_engine,
     ) as mock_engine_class:
         benchmark.run_model()
