+++ tests/stability/test_stability.py
@@ -154,7 +154,7 @@ def test_full_run_with_mocked_engine(stability_benchmark, mock_jaxmd_simulation_
     benchmark = stability_benchmark
     mock_engine = mock_jaxmd_simulation_engine()
     with patch(
-        "mlipaudit.utils.simulation.JaxMDSimulationEngine",
+        "mlip.simulation.jax_md.JaxMDSimulationEngine",
         return_value=mock_engine,
     ) as mock_engine_class:
         benchmark.run_model()
